+++ pyscriptjs/src/plugins/python/py_markdown.py
@@ -15,8 +15,8 @@ class MyPlugin(Plugin):
     def configure(self, config):
         console.log(f"configuration received: {config}")
 
-    def afterStartup(self, runtime):
-        console.log(f"runtime received: {runtime}")
+    def afterStartup(self, interpreter):
+        console.log(f"interpreter received: {interpreter}")
 
 
 plugin = MyPlugin("py-markdown")