+++ pyscriptjs/tests/py-unit/test_python_plugins.py
@@ -8,10 +8,10 @@ class TestPyMarkdown:
         console_mock = Mock()
         monkeypatch.setattr(py_markdown, "console", console_mock)
         config = "just a config"
-        runtime = "just a runtime"
+        interpreter = "just an interpreter"
 
         py_markdown.plugin.configure(config)
         console_mock.log.assert_called_with("configuration received: just a config")
 
-        py_markdown.plugin.afterStartup(runtime)
-        console_mock.log.assert_called_with("runtime received: just a runtime")
+        py_markdown.plugin.afterStartup(interpreter)
+        console_mock.log.assert_called_with("interpreter received: just an interpreter")